+++ backend/internal/service/payment_service.go
@@ -271,11 +271,17 @@ func psSliceContains(sl []string, s string) bool {
 	return false
 }
 
+// Subscription validity period unit constants.
+const (
+	validityUnitWeek  = "week"
+	validityUnitMonth = "month"
+)
+
 func psComputeValidityDays(days int, unit string) int {
 	switch unit {
-	case "week":
+	case validityUnitWeek:
 		return days * 7
-	case "month":
+	case validityUnitMonth:
 		return days * 30
 	default:
 		return days
