+++ frontend/src/i18n/locales/en.ts
@@ -308,6 +308,8 @@ export default {
     chooseFile: 'Choose File',
     notAvailable: 'N/A',
     now: 'Now',
+    today: 'Today',
+    tomorrow: 'Tomorrow',
     unknown: 'Unknown',
     minutes: 'min',
     time: {
@@ -5237,6 +5239,8 @@ export default {
       createdAt: 'Created',
       cancel: 'Cancel Order',
       userId: 'User ID',
+      orderType: 'Order Type',
+      actions: 'Actions',
       requestRefund: 'Request Refund',
     },
     result: {
